+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Compatibility:
 * Add `Pathname#/` alias to `Pathname#+` (#2178).
 * Fixed issue with large `Integer`s in `Math.log` (#2184).
 * Updated `Regexp.last_match` to support `Symbol` and `String` parameter (#2179).
+* Added support for numbered block parameters (`_1` etc).
 
 Performance: